Image Filters
During the course, we were introduced to image filters which is basically about looping through the pixels of an image and change the RGB values of these pixels to get a new filtered image. I really enjoyed it and went further to discover and test Pillow's functions (the brighten function). I also created my very own filter!